requirements.txt
asgiref==3.6.0
attrs==23.1.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.1
camelcase==0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
crashtest==0.4.1
cryptography==41.0.1
cycler==0.11.0
Deprecated==1.2.14
distlib==0.3.6
Django==4.1.7
django-crispy-forms==1.14.0
django-extensions==3.2.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
dulwich==0.21.5
filelock==3.12.0
firebase-admin==6.1.0
Flask==2.2.3
fonttools==4.38.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jwcrypto==1.5.0
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.2
matplotlib==3.7.0
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
numpy==1.24.2
oauth2client==4.1.3
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.1
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
proto-plus==1.22.2
protobuf==4.23.2
psycopg2==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.14.0
PyJWT==2.7.0
pymongo==3.11.4
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
Pyrebase4==4.7.0
pyrebase5==5.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-firebase==1.2
python-jwt==4.0.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
razorpay==1.3.0
requests==2.30.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.0.post1
simplejson==3.18.4
six==1.16.0
sqlparse==0.2.4
sympy==1.11.1
threadpoolctl==3.1.0
tk==0.1.0
tomlkit==0.11.8
trove-classifiers==2023.5.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.1
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xvfbwrapper==0.2.9
zipp==3.15.0
